Fixed axis left / right auto logic
